// frontend/contracts/core/v1/models/ProblemTeamModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProblemTeamModel = {
    problem_id: string;
    team_id: string;
    team_name: string;
    team_slug: string;
    permission: string;
    created_at: string;
};