// contracts/core/v1/models/TeamModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TeamModel = {
    id: string;
    name: string;
    organization_id: string;
    description?: string;
    created_at: string;
    updated_at: string;
};